WIP: change plugin struct
