+++ apps/plugins/nft/plugin/parse_test.go
