+++ apps/plugins/nft/types.go
@@ -1,4 +1,4 @@
-package main
+package nft
 
 import (
 	"bytes"
