+++ apps/plugins/socat/plugin/socat_test.go
